+++ src/functions/battleping.js
@@ -1,4 +1,5 @@
 const BattlePingSchema = require('../_database/models/battlePingSchema')
+const RumbleEra = require('../_database/models/rumbleEraSchema')
 const { addLog } = require('./logs')
 const mongoose = require('mongoose')
 
@@ -27,6 +28,57 @@ async function addBattlePing(channel, pingRole, reward, title, countDown, defaul
     await battlePing.save().catch(error => addLog(null, error, error.stack))
 }
 
+async function getEras() {
+    let eras = await RumbleEra.aggregate(
+        [
+            {
+                $addFields: {
+                    name_lc: {
+                        $toLower: "$name"
+                    },
+                    choice: {
+                        $concat: ["$emoji", " ", "$name"]
+                    }
+                }
+            },
+            {
+                $sort: { name_lc: 1 }
+            }
+        ]
+    )
+
+    return eras
+}
+
+async function handleAnnouncementChoices(slashCommandsToAnnounce) {
+    let eras = await getEras()
+    if (eras.length !== 0) {
+        let editChoices = eras.map(t => ({ name: t.name_lc, value: t.name }))
+        let selectChoices = eras.map(t => ({ name: t.name_lc, value: t.choice }))
+        slashCommandsToAnnounce.forEach(cmd => {
+
+            switch (cmd.name) {
+                case 'era':
+                    cmd.options.forEach(opt => {
+                        if (opt.name === 'remove') {
+                            opt.options?.forEach(subopt => {
+                                if (subopt.name === "name") subopt.choices = editChoices
+                            })
+                        }
+                    })
+                    break;
+                case 'battleping':
+                    cmd.options.forEach(opt => {
+                        if (opt.name === "era") opt.choices = selectChoices
+                    })
+                    break;
+            }
+        })
+    }
+
+    return slashCommandsToAnnounce
+}
+
 async function removeBattlePing(channel) {
     let battlePing = await getBattlePing(channel)
 
@@ -43,5 +95,7 @@ module.exports = {
     name: "pinglist",
     addBattlePing,
     removeBattlePing,
-    getBattlePing
+    getBattlePing,
+    getEras,
+    handleAnnouncementChoices
 }